+++ packages/insomnia-app/app/ui/components/dropdowns/remote-workspaces-dropdown.tsx
@@ -1,101 +1,16 @@
-import React, { FC, useCallback, useEffect, useMemo, useState } from 'react';
+import React, { FC } from 'react';
 import VCS from '../../../sync/vcs';
-import type { Project } from '../../../sync/types';
 import { Dropdown, DropdownDivider, DropdownItem, Button, Tooltip } from 'insomnia-components';
 import HelpTooltip from '../help-tooltip';
-import { showAlert } from '../modals';
 import { strings } from '../../../common/strings';
-import { useSelector } from 'react-redux';
-import { selectActiveSpace, selectAllWorkspaces } from '../../redux/selectors';
 import { isLoggedIn } from '../../../account/session';
-import { isNotNullOrUndefined } from '../../../common/misc';
-import { pullProject } from '../../../sync/vcs/pull-project';
+import { useRemoteWorkspaces } from '../../hooks/workspace';
+import { useActiveSpace } from '../../hooks/space';
 
 interface Props {
   className?: string;
   vcs?: VCS | null;
 }
-
-const useRemoteWorkspaces = (vcs?: VCS) => {
-  // Fetch from redux
-  const workspaces = useSelector(selectAllWorkspaces);
-  const activeSpace = useSelector(selectActiveSpace);
-  const spaceRemoteId = activeSpace?.remoteId || undefined;
-  const isRemoteSpace = isNotNullOrUndefined(spaceRemoteId);
-  const spaceId = activeSpace?._id;
-
-  // Local state
-  const [loading, setLoading] = useState(false);
-  const [localProjects, setLocalProjects] = useState<Project[]>([]);
-  const [remoteProjects, setRemoteProjects] = useState<Project[]>([]);
-  const [pullingProjects, setPullingProjects] = useState<Record<string, boolean>>({});
-
-  // Refresh remote spaces
-  const refresh = useCallback(async () => {
-    if (!vcs || !isLoggedIn()) {
-      return;
-    }
-
-    setLoading(true);
-    const remote = await vcs.remoteProjects(spaceRemoteId);
-    const local = await vcs.localProjects();
-    setRemoteProjects(remote);
-    setLocalProjects(local);
-    setLoading(false);
-  },
-  [spaceRemoteId, vcs]);
-
-  // Find remote spaces that haven't been pulled
-  const missingProjects = useMemo(() => remoteProjects.filter(({ id, rootDocumentId }) => {
-    const localProjectExists = localProjects.find(p => p.id === id);
-    const workspaceExists = workspaces.find(w => w._id === rootDocumentId);
-    // Mark as missing if:
-    //   - the project doesn't yet exists locally
-    //   - the project exists locally but somehow the workspace doesn't anymore
-    return !(workspaceExists && localProjectExists);
-  }), [localProjects, remoteProjects, workspaces]);
-
-  // Pull a remote space
-  const pull = useCallback(async (project: Project) => {
-    if (!vcs) {
-      throw new Error('VCS is not defined');
-    }
-
-    setPullingProjects(state => ({ ...state, [project.id]: true }));
-
-    try {
-      // Clone old VCS so we don't mess anything up while working on other projects
-      const newVCS = vcs.newInstance();
-      // Remove all projects for workspace first
-      await newVCS.removeProjectsForRoot(project.rootDocumentId);
-
-      await pullProject({ vcs: newVCS, project, spaceId });
-
-      await refresh();
-    } catch (err) {
-      showAlert({
-        title: 'Pull Error',
-        message: `Failed to pull workspace. ${err.message}`,
-      });
-    } finally {
-      setPullingProjects(state => ({ ...state, [project.id]: false }));
-    }
-  }, [vcs, refresh, spaceId]);
-
-  // If the refresh callback changes, refresh
-  useEffect(() => {
-    (async () => { await refresh(); })();
-  }, [refresh]);
-
-  return {
-    isRemoteSpace,
-    loading,
-    missingProjects,
-    pullingProjects,
-    refresh,
-    pull,
-  };
-};
 
 const PullButton: FC<{disabled?: boolean, className?: string}> = ({ disabled, className }) => (
   <Button className={className} disabled={disabled}>
@@ -106,7 +21,6 @@ const PullButton: FC<{disabled?: boolean, className?: string}> = ({ disabled, cl
 
 export const RemoteWorkspacesDropdown: FC<Props> = ({ className, vcs }) => {
   const {
-    isRemoteSpace,
     loading,
     refresh,
     missingProjects,
@@ -114,7 +28,9 @@ export const RemoteWorkspacesDropdown: FC<Props> = ({ className, vcs }) => {
     pull,
   } = useRemoteWorkspaces(vcs || undefined);
 
-  // Don't show the pull dropdown if this is not a remote space
+  const { isRemoteSpace } = useActiveSpace();
+
+  // Don't show the pull dropdown if we are not in a remote space
   if (!isRemoteSpace) {
     return null;
   }
